refactor: move vllm multimodal examples into component
