+++ examples/basics/kubernetes/shared_frontend/shared_frontend.yaml
@@ -74,12 +74,12 @@ spec:
       extraPodSpec:
         mainContainer:
           image: nvcr.io/nvidia/ai-dynamo/vllm-runtime:0.5.0
-          workingDir: /workspace/examples/multimodal
+          workingDir: /workspace/examples/backends/vllm
           command:
             - /bin/sh
             - -c
           args:
-            - python3 components/encode_worker.py --model Qwen/Qwen2.5-VL-7B-Instruct
+            - python -m dynamo.vllm --multimodal-encode-worker --enable-multimodal --model Qwen/Qwen2.5-VL-7B-Instruct
     VLMWorker:
       volumeMounts:
       - name: dynamo-model-cache
@@ -93,12 +93,12 @@ spec:
       extraPodSpec:
         mainContainer:
           image: nvcr.io/nvidia/ai-dynamo/vllm-runtime:0.5.0
-          workingDir: /workspace/examples/multimodal
+          workingDir: /workspace/examples/backends/vllm
           command:
             - /bin/sh
             - -c
           args:
-            - python3 components/worker.py --model Qwen/Qwen2.5-VL-7B-Instruct --worker-type prefill
+            - python -m dynamo.vllm --multimodal-worker --enable-multimodal --enable-mm-embeds --model Qwen/Qwen2.5-VL-7B-Instruct
     Processor:
       volumeMounts:
       - name: dynamo-model-cache
@@ -112,9 +112,9 @@ spec:
       extraPodSpec:
         mainContainer:
           image: nvcr.io/nvidia/ai-dynamo/vllm-runtime:0.5.0
-          workingDir: /workspace/examples/multimodal
+          workingDir: /workspace/examples/backends/vllm
           command:
             - /bin/sh
             - -c
           args:
-            - 'python3 components/processor.py --model Qwen/Qwen2.5-VL-7B-Instruct --prompt-template "<|im_start|>system\nYou are a helpful assistant.<|im_end|>\n<|im_start|>user\n<|vision_start|><|image_pad|><|vision_end|><prompt><|im_end|>\n<|im_start|>assistant\n"'
+            - python -m dynamo.vllm --multimodal-processor --enable-multimodal --model Qwen/Qwen2.5-VL-7B-Instruct
